[20241215] Update `BigQueryConfig` functions ## Updates - Moved `Dataset` and `Source` to the constants package. - Moved `getDatasetId` & `getFullyQualifiedTableName` functions to `BigQueryConfig` class and refactored their usage in the `etl-stream` module.
